requirements.txt
certifi==2024.6.2
charset-normalizer==3.3.2
filelock==3.14.0
fsspec==2024.6.0
idna==3.7
Jinja2==3.1.4
MarkupSafe==2.1.5
mlab-py==1.0.12
mpmath==1.3.0
networkx==3.3
numpy==1.26.4
pandas==2.2.2
pillow==10.3.0
python-dateutil==2.9.0.post0
pytz==2024.1
requests==2.32.3
six==1.16.0
sympy==1.12.1
torch==2.2.2
torchvision==0.17.2
typing_extensions==4.12.1
tzdata==2024.1
urllib3==2.2.1
